Add files via upload
This is my attempt at Q2! Whoever is the code checker can check it and let me know what they think. I did not construct all the plots yet, but I made the function to do so. All you have to do is call the function, and specify n and the parameters (in this case, shape = 1, scale = 2).